requirements.txt
streamlit==1.38.0
python-dotenv
requests

llama-index
llama-index-core==0.13.0
llama-index-llms-ollama

google-generativeai
google-genai
python-dotenv
llama-index-llms-gemini

PyMuPDF